[IMP] delivery_package_number: Ask only once for the number of packages wizard. TT48170
