+++ delivery_package_number/README.rst
@@ -2,7 +2,7 @@
 Stock Picking Package Number
 ============================
 
-.. 
+..
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
@@ -82,8 +82,8 @@ Contributors
 
   * Pedro M. Baeza
   * David Vidal
-  * Marçal Isern
   * Carlos Roca
+  * Sergio Teruel
 
 
  * `Sygel <https://www.sygel.es>`_:
