+++ delivery_package_number/static/description/index.html
@@ -425,8 +425,8 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <li><a class="reference external" href="https://www.tecnativa.com">Tecnativa</a>:<ul>
 <li>Pedro M. Baeza</li>
 <li>David Vidal</li>
-<li>Marçal Isern</li>
 <li>Carlos Roca</li>
+<li>Sergio Teruel</li>
 </ul>
 </li>
 </ul>
