+++ delivery_package_number/tests/test_delivery_package_number.py
@@ -66,7 +66,9 @@ class TestDeliveryPackageNumber(TransactionCase):
         picking = order.picking_ids
         picking.move_ids.quantity_done = 2
         picking.number_of_packages = 2
-        action = picking.button_validate()
+        action = picking.with_context(
+            test_delivery_package_number=True
+        ).button_validate()
         backorder_wizard = Form(
             self.env[action["res_model"]].with_context(**action["context"])
         ).save()
