+++ delivery_package_number/wizard/__init__.py
@@ -1,4 +1,3 @@
-from . import stock_number_package_mixin
 from . import stock_inmediate_transfer
 from . import stock_backorder_confirmation
 from . import stock_number_package_validate_wiz
