--- delivery_package_number/wizard/stock_backorder_confirmation_views.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="delivery_package_number_view_backorder_confirmation" model="ir.ui.view">
-        <field name="name">delivery.package.number.view.backorder.confirmation</field>
-        <field name="model">stock.backorder.confirmation</field>
-        <field name="inherit_id" ref="stock.view_backorder_confirmation" />
-        <field name="arch" type="xml">
-            <xpath expr="//footer" position="before">
-                <field name="ask_number_of_packages" invisible="1" />
-                <group attrs="{'invisible': [('ask_number_of_packages','=',False)]}">
-                    <field name="number_of_packages" />
-                </group>
-            </xpath>
-            <xpath expr="//button[@special='cancel']" position="after">
-                <field
-                    name="print_package_label"
-                    widget="boolean_toggle"
-                    attrs="{'invisible': [('ask_number_of_packages','=',False)]}"
-                />
-                <label
-                    for="print_package_label"
-                    attrs="{'invisible': [('ask_number_of_packages','=',False)]}"
-                />
-            </xpath>
-        </field>
-    </record>
-</odoo>
